fix(physics): player cannot jump when standing on CBox (#122)
there was no collision event raised for CBox:es and there was no
handler for it (see changes in InputSystem.cs). this commit fixes it
by handling the collision event in the same manner the the
"collisionwithground" event has always been handled.

NOTE: there are other issues with the way collision event handling is
done in the input system - see checks for playersInt.Count == 0 and
what not - it's not a sane solution to recheck for all players *every
single time* a collision occurs. that, however, is a separate issue
that potentially needs to be fixed.

See #122